Add app tests
