+++ bdd_jx_suite_test.go
@@ -2,6 +2,7 @@ package bdd_jx
 
 import (
 	"encoding/json"
+	"fmt"
 	"github.com/jenkins-x/jx/pkg/util"
 	"io/ioutil"
 	"log"
@@ -20,8 +21,12 @@ import (
 )
 
 var ReporterTestGrid *reporters.ReporterTestGrid
-
+func init() {
+	fmt.Println("initfunc")
+}
 func TestBddJx(t *testing.T) {
+	fmt.Println("TestBddJxStart")
+
 	specFailures := make(map[string][]bool)
 	reps := []Reporter{}
 	ReporterTestGrid = &reporters.ReporterTestGrid{
@@ -40,6 +45,7 @@ func TestBddJx(t *testing.T) {
 
 var _ = BeforeSuite(func() {
 	var err error
+	fmt.Println("BeforeSuiteStuff")
 	WorkDir, err = ioutil.TempDir("", TempDirPrefix)
 	Expect(err).NotTo(HaveOccurred())
 	err = os.MkdirAll(WorkDir, 0760)
