+++ main.go
@@ -27,7 +27,9 @@ import (
 )
 
 func main() { /* usual main func */ }
-
+func init() {
+	fmt.Println("testline");
+}
 var (
 	// TempDirPrefix The prefix to append to applicationss created in testing
 	TempDirPrefix = "bdd-"
@@ -352,30 +354,9 @@ func (t *Test) ExpectUrlReturns(url string, expectedStatusCode int, maxDuration 
 }
 
 // AddAppTests Creates a jx add app test
-func AppTests() []bool {
-	flag.Parse()
-	includedApps := *IncludeApps
-	if includedApps != "" {
-		includedAppList := strings.Split(strings.TrimSpace(includedApps), ",")
-		tests := make([]bool, len(includedAppList))
-		for _, testAppName := range includedAppList {
-			nameAndVersion := strings.Split(testAppName, ":")
-			if len(nameAndVersion) == 2 {
-				tests = append(tests, AppTest(nameAndVersion[0], nameAndVersion[1]))
-			} else {
-				tests = append(tests, AppTest(testAppName, ""))
-			}
-
-		}
-		return tests
-	} else {
-		return nil
-	}
-}
 
 // AddAppTests Creates a jx add app test
 func AllQuickstartsTest() []bool {
-	flag.Parse()
 	includedQuickstarts := *IncludeQuickstarts
 	if includedQuickstarts != "" {
 		includedQuickstartList := strings.Split(strings.TrimSpace(includedQuickstarts), ",")
@@ -389,75 +370,7 @@ func AllQuickstartsTest() []bool {
 	}
 }
 
-func AppTest(testAppName string, version string) bool {
-	return Describe("test app "+testAppName+"\n", func() {
-		var T Test
 
-		BeforeEach(func() {
-			T = Test{
-				ApplicationName: TempDirPrefix + testAppName + "-" + strconv.FormatInt(GinkgoRandomSeed(), 10),
-				WorkDir:         WorkDir,
-				Factory:         clients.NewFactory(),
-			}
-			T.GitProviderURL()
-		})
-
-		_ = T.AddAppTests(testAppName, version)
-		_ = T.DeleteAppTests(testAppName)
-
-	})
-}
-
-// AddAppTests add app tests
-func (t *Test) AddAppTests(testAppName string, version string) bool {
-	return Describe("Given valid parameters", func() {
-		Context("when running jx add app "+testAppName, func() {
-			commandTimeout := 1 * time.Hour
-			helmAppName := testAppName + "-" + testAppName
-			It("Ensure the app is added\n", func() {
-				By("The App resource does not exist before creation\n")
-				c := "kubectl"
-				args := []string{"get", "app", helmAppName}
-				t.ExpectCommandExecution(t.WorkDir, commandTimeout, 1, c, args...)
-				By("Add app exits with signal 0\n")
-				c = "jx"
-				args = []string{"add", "app", testAppName, "--repository", DefaultRepositoryURL}
-				if version != "" {
-					args = append(args, "--version", version)
-				}
-				t.ExpectCommandExecution(t.WorkDir, commandTimeout, 0, c, args...)
-				By("The App resource exists after creation\n")
-				c = "kubectl"
-				args = []string{"get", "app", helmAppName}
-				t.ExpectCommandExecution(t.WorkDir, commandTimeout, 0, c, args...)
-			})
-		})
-	})
-}
-
-// DeleteAppTests delete app tests
-func (t *Test) DeleteAppTests(testAppName string) bool {
-	return Describe("Given valid parameters", func() {
-		Context("when running jx delete app "+testAppName, func() {
-			commandTimeout := 1 * time.Hour
-			helmAppName := testAppName + "-" + testAppName
-			It("Ensure it is deleted\n", func() {
-				By("The App resource exists before deletion\n")
-				c := "kubectl"
-				args := []string{"get", "app", helmAppName}
-				t.ExpectCommandExecution(t.WorkDir, commandTimeout, 0, c, args...)
-				By("Delete app exits with signal 0\n")
-				c = "jx"
-				args = []string{"delete", "app", testAppName}
-				t.ExpectCommandExecution(t.WorkDir, commandTimeout, 0, c, args...)
-				By("The App resource was removed\n")
-				c = "kubectl"
-				args = []string{"get", "app", helmAppName}
-				t.ExpectCommandExecution(t.WorkDir, commandTimeout, 1, c, args...)
-			})
-		})
-	})
-}
 
 //CreateQuickstartTest creates a quickstart test for the given quickstart
 func CreateQuickstartTest(quickstartName string) bool {
